requirements.txt
flask
openai
pandas
python-dotenv
requests
langchain
numpy
scikit-learn
transformers
torch
faiss-cpu
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
python-jose
fastapi
uvicorn